Actualizado

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
     $table->id();
-    $table->integer('DNI')->unique(); // Asegúrate que no tenga ->nullable()
+    $table->unsignedBigInteger('DNI')->unique(); // Asegúrate que no tenga ->nullable()
     $table->string('name');
     $table->string('apellido_paterno');
     $table->string('apellido_materno');
